Fix build error on macOS (#1529)
This is the same as #1293, but on newly added code.

+++ stablehlo/reference/Element.cpp
@@ -623,7 +623,8 @@ Element complex(const Element &e1, const Element &e2) {
 
 Element convert(Type type, bool value) {
   if (isSupportedBooleanType(type)) return Element(type, value);
-  return convert(type, value ? 1UL : 0UL);
+  return convert(type,
+                 value ? static_cast<uint64_t>(1) : static_cast<uint64_t>(0));
 }
 
 Element convert(Type type, APSInt value) {
@@ -813,7 +814,7 @@ Element power(const Element &e1, const Element &e2) {
       if (base.abs().isOne())
         exponent = exponent.abs();
       else
-        return convert(type, 0L);
+        return convert(type, static_cast<int64_t>(0));
     }
     APInt result(base.getBitWidth(), 1, isSigned);
     while (!exponent.isZero()) {
@@ -901,9 +902,9 @@ Element sign(const Element &el) {
 
   if (isSupportedIntegerType(type)) {
     auto elVal = el.getIntegerValue();
-    if (elVal.isNegative()) return convert(type, -1L);
-    if (elVal.isZero()) return convert(type, 0L);
-    return convert(type, 1L);
+    if (elVal.isNegative()) return convert(type, static_cast<int64_t>(-1));
+    if (elVal.isZero()) return convert(type, static_cast<int64_t>(0));
+    return convert(type, static_cast<int64_t>(1));
   }
 
   if (isSupportedFloatType(type)) {
